requirements.txt
Django==3.2
django-filter==23.1
djangorestframework==3.12.4
djoser==2.1.0
requests==2.26.0